requirements.txt
Django>=5.2,<6.0
djangorestframework
gunicorn
